final_implementation/requirements.txt
# Final Implementation Requirements
# Minimal, proven dependencies

# Core web automation
playwright==1.40.0
beautifulsoup4==4.12.2

# Async support
aiohttp==3.9.1
aiofiles==23.2.1

# Gmail integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Environment management
python-dotenv==1.0.0

# Data processing (optional but useful)
pandas==2.1.4
openpyxl==3.1.2